refactor: inject demo traces
