+++ apps/ui/src/pages/router.tsx
@@ -1,7 +1,6 @@
 import { HashRouter, Route, Routes } from "react-router-dom";
 
 import { Dashboard } from "./dashboard";
-import { Demo } from "./demo/demo";
 import { Schema } from "./schema";
 
 export function Router() {
@@ -10,7 +9,6 @@ export function Router() {
       <Routes>
         <Route path="/" element={<Dashboard />} />
         <Route path="*" element={<Dashboard />} />
-        <Route path="/demo" element={<Demo />} />
         <Route path="/schema/:schemaId" element={<Schema />} />
         <Route path="/schema/:schemaId/trace/:traceId" element={<Schema />} />
       </Routes>
